requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
PyMuPDF
langchain-community
pypdf